refactor : service 단계에서 예외 처리하기 개선
