+++ src/main/java/com/example/soso/security/domain/CustomUserDetailsService.java
@@ -1,5 +1,7 @@
 package com.example.soso.security.domain;
 
+import com.example.soso.global.exception.domain.UserErrorCode;
+import com.example.soso.global.exception.util.UserAuthException;
 import com.example.soso.users.repository.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -17,7 +19,7 @@ public class CustomUserDetailsService implements UserDetailsService {
     public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
 
         if (userId == null || userId.isBlank() || !userRepository.existsById(userId)) {
-            throw new UsernameNotFoundException("유효하지 않은 사용자 ID입니다.");
+            throw new UserAuthException(UserErrorCode.UNAUTHORIZED_USER);
         }
 
         return new CustomUserDetails(userId);
