+++ src/main/java/com/example/soso/users/service/AuthServiceImpl.java
@@ -1,6 +1,7 @@
-package com.example.soso.users.controller;
+package com.example.soso.users.service;
 
 import com.example.soso.global.config.CookieUtil;
+import com.example.soso.global.exception.util.InvalidTokenException;
 import com.example.soso.jwt.JwtProperties;
 import com.example.soso.jwt.JwtProvider;
 import com.example.soso.jwt.JwtTokenDto;
@@ -8,6 +9,8 @@ import com.example.soso.jwt.RefreshTokenRedisService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import static com.example.soso.global.exception.domain.TokenErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -17,27 +20,27 @@ public class AuthServiceImpl implements AuthService {
     private final RefreshTokenRedisService refreshTokenService;
     private final JwtProperties jwtProperties;
 
+    @Override
     public JwtTokenDto refreshAccessToken(String refreshToken, HttpServletResponse response) {
-
         if (!jwtProvider.validateToken(refreshToken)) {
-            throw new IllegalArgumentException("유효하지 않은 토큰입니다.");
+            throw new InvalidTokenException(INVALID_TOKEN);
         }
 
         String userId = jwtProvider.getSubject(refreshToken);
-        String stored = refreshTokenService.get(userId);
+        String storedToken = refreshTokenService.get(userId);
 
-        if (!refreshToken.equals(stored)) {
-            throw new IllegalArgumentException("재사용되었거나 탈취된 토큰입니다.");
+        if (!refreshToken.equals(storedToken)) {
+            throw new InvalidTokenException(REFRESH_TOKEN_NOT_FOUND);
         }
 
-        // RTR: refreshToken 교체
+        // 1. RTR: refreshToken 교체 및 재저장
         String newRefreshToken = jwtProvider.generateRefreshToken(userId);
         refreshTokenService.save(userId, newRefreshToken, jwtProperties.getRefreshTokenValidityInMs());
 
-        // 쿠키에 다시 설정
+        // 2. 새 refreshToken 쿠키에 저장
         CookieUtil.addRefreshTokenCookie(response, newRefreshToken, jwtProperties.getRefreshTokenValidityInMs());
 
-        // accessToken 발급
+        // 3. 새 accessToken 발급
         String newAccessToken = jwtProvider.generateAccessToken(userId);
 
         return new JwtTokenDto(newAccessToken);
